+++ test/Loop.java
@@ -11,9 +11,9 @@ public class Loop {
         int t = 0;
         while(window.isOpen()) {
             double time = t/10.0;
-            int amplitude = (int) (sin(time) * 50 + 50);
+            double amplitude = sin(time) * 50 + 50;
             window.setColor((int) (sin(time) * 255), (int) (sin(time/PI) * 255), 10);
-            window.fillRect(amplitude, (int) (cos(time) * 50 + 50), amplitude + 10, amplitude + 10);
+            window.fillRect(amplitude, cos(time) * 50 + 50, amplitude + 10, amplitude + 10);
             window.refresh(20);
             t++;
         }
